Cleaned up the init api
