+++ src/main/java/com/falsepattern/lumina/internal/data/ChunkNBTManager.java
@@ -23,10 +23,9 @@ package com.falsepattern.lumina.internal.data;
 
 import com.falsepattern.chunk.api.ChunkDataManager;
 import com.falsepattern.chunk.api.ChunkDataRegistry;
+import com.falsepattern.lumina.api.LumiChunkAPI;
 import com.falsepattern.lumina.api.chunk.LumiChunk;
-import com.falsepattern.lumina.api.chunk.loading.LumiChunkHelper;
 import com.falsepattern.lumina.api.lighting.LumiLightingEngine;
-import com.falsepattern.lumina.api.chunk.loading.LuminaChunkTaskQueue;
 import lombok.NoArgsConstructor;
 import lombok.val;
 import lombok.var;
@@ -118,7 +117,7 @@ public final class ChunkNBTManager implements ChunkDataManager.ChunkNBTDataManag
             val worldTag = input.getCompoundTag(worldTagName);
             val worldProviderVersion = worldProvider.worldProviderVersion();
             if (!worldProviderVersion.equals(worldTag.getString(WORLD_PROVIDER_VERSION_NBT_TAG_NAME))) {
-                LumiChunkHelper.scheduleLightingEngineReinitialization(chunk, lightingEngine);
+                LumiChunkAPI.scheduleChunkLightingEngineInit(chunk);
                 continue;
             }
