+++ src/main/java/com/falsepattern/lumina/internal/mixin/plugin/Mixin.java
@@ -58,6 +58,7 @@ public enum Mixin implements IMixin {
 
     // region LUMINA Initialization
     common_init_LumiChunkBaseInitImplMixin(COMMON, always(), "init.LumiChunkBaseInitImplMixin"),
+    common_init_LumiChunkInitTaskQueueImplMixin(COMMON, always(), "init.LumiChunkInitTaskQueueImplMixin"),
 
     common_init_LumiSubChunkBaseInitImplMixin(COMMON, always(), "init.LumiSubChunkBaseInitImplMixin"),
